+++ waves/tests/test_latinhypercube.py
@@ -14,55 +14,6 @@ from waves._settings import _hash_coordinate_key, _set_coordinate_key
 class TestLatinHypercube:
     """Class for testing LatinHypercube parameter study generator class"""
 
-    validate_input = {
-        "good schema": (
-            {'num_simulations': 1, 'parameter_1': {'distribution': 'norm', 'kwarg1': 1}},
-            does_not_raise()
-        ),
-        "not a dict": (
-            'not a dict',
-            pytest.raises(TypeError)
-        ),
-        "missing num_simulation": (
-            {},
-            pytest.raises(AttributeError)
-        ),
-        "num_simulation non-integer": (
-            {'num_simulations': 'not_a_number'},
-            pytest.raises(TypeError)
-        ),
-        "missing distribution": (
-            {'num_simulations': 1, 'parameter_1': {}},
-            pytest.raises(AttributeError)
-        ),
-        "distribution non-string": (
-            {'num_simulations': 1, 'parameter_1': {'distribution': 1}},
-            pytest.raises(TypeError)
-        ),
-        "distribution bad identifier": (
-            {'num_simulations': 1, 'parameter_1': {'distribution': 'my norm'}},
-            pytest.raises(TypeError)
-        ),
-        "kwarg bad identifier": (
-            {'num_simulations': 1, 'parameter_1': {'distribution': 'norm', 'kwarg 1': 1}},
-            pytest.raises(TypeError)
-        )
-    }
-
-    @pytest.mark.unittest
-    @pytest.mark.parametrize('parameter_schema, outcome',
-                             validate_input.values(),
-                             ids=validate_input.keys())
-    def test__validate(self, parameter_schema, outcome):
-        with patch("waves.parameter_generators.LatinHypercube._generate_parameter_distributions") as mock_distros, \
-             outcome:
-            try:
-                # Validate is called in __init__. Do not need to call explicitly.
-                TestValidate = LatinHypercube(parameter_schema)
-                mock_distros.assert_called_once()
-            finally:
-                pass
-
     generate_input_interface = "parameter_schema, random_state, expected_samples, expected_quantiles, " \
                                "expected_scipy_kwds"
     generate_input = {
@@ -112,29 +63,18 @@ class TestLatinHypercube:
                       expected_samples, expected_quantiles, expected_scipy_kwds):
         parameter_names = [key for key in parameter_schema.keys() if key != 'num_simulations']
         TestGenerate = LatinHypercube(parameter_schema)
-        TestGenerate.generate(lhs_kwargs={'random_state': random_state})
+        TestGenerate.generate(kwargs={'random_state': random_state})
         samples_array = TestGenerate._samples
         quantiles_array = TestGenerate._quantiles
         assert numpy.allclose(samples_array, expected_samples)
         assert numpy.allclose(quantiles_array, expected_quantiles)
         # Verify that the parameter set name creation method was called
         expected_set_names = [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
-        assert list(TestGenerate._parameter_set_names.values()) == expected_set_names 
+        assert list(TestGenerate._parameter_set_names.values()) == expected_set_names
         # Check that the parameter set names are correctly populated in the parameter study Xarray Dataset
         expected_set_names = [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
         parameter_set_names = list(TestGenerate.parameter_study[_set_coordinate_key])
         assert numpy.all(parameter_set_names == expected_set_names)
-
-    @pytest.mark.unittest
-    @pytest.mark.parametrize(generate_input_interface,
-                             generate_input.values(),
-                             ids=generate_input.keys())
-    def test_generate_parameter_distributions(self, parameter_schema, random_state,
-                                              expected_samples, expected_quantiles, expected_scipy_kwds):
-        TestDistributions = LatinHypercube(parameter_schema)
-        assert TestDistributions._parameter_names == list(TestDistributions.parameter_distributions.keys())
-        for parameter_name, expected_kwds in zip(TestDistributions._parameter_names, expected_scipy_kwds):
-            assert TestDistributions.parameter_distributions[parameter_name].kwds == expected_kwds
 
     merge_test = {
         'increase simulations': (
@@ -160,10 +100,10 @@ class TestLatinHypercube:
                              ids=merge_test.keys())
     def test_merge(self, first_schema, second_schema, random_state, expected_samples, expected_quantiles):
         TestMerge1 = LatinHypercube(first_schema)
-        TestMerge1.generate(lhs_kwargs={'random_state': random_state})
+        TestMerge1.generate(kwargs={'random_state': random_state})
         with patch('xarray.open_dataset', return_value=TestMerge1.parameter_study):
             TestMerge2 = LatinHypercube(second_schema, previous_parameter_study='dummy_string')
-            TestMerge2.generate(lhs_kwargs={'random_state': random_state})
+            TestMerge2.generate(kwargs={'random_state': random_state})
         samples = TestMerge2._samples.astype(float)
         quantiles = TestMerge2._quantiles.astype(float)
         assert numpy.allclose(samples, expected_samples)